requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
Deprecated==1.2.14
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isodate==0.6.1
joblib==1.3.2
msrest==0.7.1
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
replicate==0.23.1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vonage==3.13.0
vonage-jwt==1.1.0
wrapt==1.16.0
yarl==1.9.4
